+++ R/cocons.R
@@ -30,7 +30,7 @@
 #' \item \code{"smooth.limits"}: (\code{numeric vector}) specifying the allowed range of variation for the spatially varying smoothness.
 #' \item \code{"taper"}: (\code{numeric}) specifying the desired taper function from the spam package (only for "sparse" coco objects).
 #' \item \code{"delta"}: (\code{numeric}) specifying the taper range/scale (only for "sparse" coco objects).
-#' \item \code{"skip.scale"}: (\code{integer vector}) index of those variables in \code{data} that should not be scaled during the optimization.
+#' \item \code{"skip.scale"}: (\code{integer vector}) By default, all covariates are scaled. \code{skip.scale} allows to specify the index of those variables in \code{data} that should not be scaled during the optimization.
 #' }
 #' 
 #' @usage coco(type, data, locs, z, model.list, info, output = list())
@@ -92,6 +92,10 @@ coco <- function(type,
   
   if(is.vector(z)){
     z <- matrix(z, ncol = 1)
+  }
+  
+  if(is.data.frame(locs)){
+    locs <- as.matrix(locs)
   }
   
   .cocons.check.z(z, data)